+++ packages/stream_chat_persistence/lib/src/dao/channel_query_dao.dart
@@ -45,13 +45,12 @@ class ChannelQueryDao extends DatabaseAccessor<MoorChatDatabase>
         }
 
         await batch((it) {
-          it.insertAll(
+          it.insertAllOnConflictUpdate(
             channelQueries,
             cids
                 .map((cid) =>
                     ChannelQueryEntity(queryHash: hash, channelCid: cid))
                 .toList(),
-            mode: InsertMode.insertOrReplace,
           );
         });
       });