+++ packages/stream_chat_persistence/lib/src/entity/messages.dart
@@ -77,8 +77,8 @@ class Messages extends Table {
   TextColumn get pinnedByUserId => text().nullable()();
 
   /// The channel cid of which this message is part of
-  TextColumn get channelCid => text().nullable().customConstraint(
-      'NULLABLE REFERENCES channels(cid) ON DELETE CASCADE')();
+  TextColumn get channelCid =>
+      text().customConstraint('REFERENCES channels(cid) ON DELETE CASCADE')();
 
   /// A Map of [messageText] translations.
   TextColumn get i18n => text().nullable().map(MapConverter<String>())();