+++ packages/stream_chat/CHANGELOG.md
@@ -1,8 +1,20 @@
 ## Upcoming
 
+🛑️ Breaking Changes from `2.2.1`
+
+- Added 6 new methods in `ChatPersistenceClient`.
+    - `bulkUpdateMessages`
+    - `bulkUpdatePinnedMessages`
+    - `bulkUpdateMembers`
+    - `bulkUpdateReads`
+    - `updatePinnedMessageReactions`
+    - `deletePinnedMessageReactionsByMessageId`
+
 ✅ Added
 
-- Add support for `next`, `previous` value pagination in `client.search`, [read more.](https://getstream.io/chat/docs/other-rest/search/#pagination)
+- Added `Filter.contains` and `Filter.empty`
+- Added support for `next`, `previous` value pagination in `client.search`
+  , [read more.](https://getstream.io/chat/docs/other-rest/search/#pagination)
 
 ## 2.2.1